Move McpSesionHost module to private folder _host
Fixes duplicate docstring error
