+++ python/packages/autogen-ext/src/autogen_ext/tools/mcp/__init__.py
@@ -1,12 +1,12 @@
 from ._actor import McpSessionActor
 from ._config import McpServerParams, SseServerParams, StdioServerParams, StreamableHttpServerParams
 from ._factory import mcp_server_tools
+from ._host import ChatCompletionClientElicitor, Elicitor, GroupChatAgentElicitor, McpSessionHost
 from ._session import create_mcp_server_session
 from ._sse import SseMcpToolAdapter
 from ._stdio import StdioMcpToolAdapter
 from ._streamable_http import StreamableHttpMcpToolAdapter
 from ._workbench import McpWorkbench
-from .host import ChatCompletionClientElicitor, Elicitor, GroupChatAgentElicitor, McpSessionHost
 
 __all__ = [
     "create_mcp_server_session",
